Split v24_2 prediction and prediction_studio into multi-module packages
The v24_2 `prediction.py` (~600 lines) and `prediction_studio.py`
(~535 lines) modules were the largest remaining single files in the
versioned API client. Split each along its natural seams (mixin / sync
class / async class), following the leading-underscore internal-module
convention used elsewhere in the v5 sweep.

Layout per package:
- `__init__.py`  — re-exports the public surface
- `_mixin.py`    — `_PredictionV24_2Mixin` / `_PredictionStudioV24_2Mixin`
- `_sync.py`     — concrete sync class
- `_async.py`    — concrete async class

Pure mechanical split: no public API, signature, or behaviour changes.
External imports (`from ...v24_2.prediction import Prediction`, etc.)
continue to work unchanged.
