+++ python/pdstools/infinity/resources/prediction_studio/v24_2/prediction/__init__.py
@@ -1,0 +1,9 @@
+from ._async import AsyncPrediction
+from ._mixin import _PredictionV24_2Mixin
+from ._sync import Prediction
+
+__all__ = [
+    "AsyncPrediction",
+    "Prediction",
+    "_PredictionV24_2Mixin",
+]
