+++ python/pdstools/infinity/resources/prediction_studio/v24_2/prediction_studio/__init__.py
@@ -1,0 +1,9 @@
+from ._async import AsyncPredictionStudio
+from ._mixin import _PredictionStudioV24_2Mixin
+from ._sync import PredictionStudio
+
+__all__ = [
+    "AsyncPredictionStudio",
+    "PredictionStudio",
+    "_PredictionStudioV24_2Mixin",
+]
